+++ src/subdomains/core/buy-crypto/process/services/__tests__/buy-crypto.service.spec.ts
@@ -8,9 +8,9 @@ import { CryptoRouteService } from 'src/subdomains/core/buy-crypto/routes/crypto
 import { createCustomHistory } from 'src/subdomains/core/history/dto/__mocks__/history.dto.mock';
 import { BuyFiatService } from 'src/subdomains/core/sell-crypto/process/services/buy-fiat.service';
 import { BankDataService } from 'src/subdomains/generic/user/models/bank-data/bank-data.service';
-import { MultiAccountIbanService } from 'src/subdomains/generic/user/models/multi-account-iban/multi-account-iban.service';
 import { UserService } from 'src/subdomains/generic/user/models/user/user.service';
 import { BankTxService } from 'src/subdomains/supporting/bank-tx/bank-tx/bank-tx.service';
+import { MultiAccountIbanService } from 'src/subdomains/supporting/bank/multi-account-iban/multi-account-iban.service';
 import { createCustomCryptoInput } from 'src/subdomains/supporting/payin/entities/__mocks__/crypto-input.entity.mock';
 import { TransactionRequestService } from 'src/subdomains/supporting/payment/services/transaction-request.service';
 import { BuyRepository } from '../../../routes/buy/buy.repository';
